ModelsViewset agregated

+++ movies/management/commands/download.py
@@ -29,7 +29,10 @@ class Command(BaseCommand):
             url="http://www.omdbapi.com/?s={}&apikey=c3940747".format(title)
             results = requests.get(url)
             movie = results.json()
-            movie= movie['Search']
+            try:
+             movie= movie['Search']
+            except:
+               exit()
             ids=[]
             for i in movie:
                 ids.append(i['imdbID'])
@@ -69,37 +72,44 @@ class Command(BaseCommand):
                       director=director.split(',')
                       director=director[0]
                       actors=results['Actors']
-                      actors=actors.split()
+                      actors=actors.split(',')
                       actors=actors[0]
                       if date_release=='N/A':
                           date_release = datetime.datetime.strptime('01 jun 1999', '%d %b %Y')
 
                       else:
                           date_release = datetime.datetime.strptime(date_release, '%d %b %Y')
-
+                      exist=False
                       if(ext=="N/A"):
                           pass
                       else:
                           try:
                            os.mkdir(settings.MEDIA_ROOT+"/movie/"+movie_title)
                           except:
-                              print("la pelicula ya existe")
-                          poster_name=settings.MEDIA_ROOT+"/movie/"+movie_title+"/"+movie_title+"."+ext
-                          f = open(poster_name, "wb")
-                          try:
-                            response = requests.get(response)
-                            f.write(response.content)
-                          except:
-                            poster_name="prueba"
-                          f.close()
-                          poster_name="movie/"+movie_title+"/"+movie_title+"."+ext
-                          actors=Actor.objects.create(name=actors,age=0)
-                          director=MovieDirector.objects.create(name=director,age=0)
-                          Movie.objects.create(title=movie_title, duration=duration, poster=poster_name, detail=detail,
-                                               genre=genre, original_language=language,
-                                               country=country, release_date=date_release, year=year,
-                                               classification='N/A', trailer_url=trailer,
-                                               directors=director, actors=actors, rating=0.0)
+                              exist=True
+                          if exist==False:
+                              poster_name=settings.MEDIA_ROOT+"/movie/"+movie_title+"/"+movie_title+"."+ext
+                              f = open(poster_name, "wb")
+                              try:
+                                response = requests.get(response)
+                                f.write(response.content)
+                              except:
+                                poster_name="prueba"
+                              f.close()
+                              poster_name="movie/"+movie_title+"/"+movie_title+"."+ext
+                              try:
+                                  actors=Actor.objects.get(name__icontains=actors)
+                              except:
+                                  actors=Actor.objects.create(name=actors,age=0)
+                              try:
+                                  director = MovieDirector.objects.get(name__icontains=director)
+                              except:
+                                  director=MovieDirector.objects.create(name=director,age=0)
+                              Movie.objects.create(title=movie_title, duration=duration, poster=poster_name, detail=detail,
+                                                   genre=genre, original_language=language,
+                                                   country=country, release_date=date_release, year=year,
+                                                   classification='N/A', trailer_url=trailer,
+                                                   directors=director, actors=actors, rating=0.0)
 
 
         print("success")
